@@ -3,20 +3,21 @@
   "version": "5.6.3",
   "description": "Designed for Subtitles. Make watching  a great experience.",
   "author": "Stefan Breitenstein",
+  "type": "module",
   "scripts": {
     "eslint": "eslint . --ext .ts",
     "start:chrome": "webpack --env browser=chrome --watch --env add_manifest_key=true",
     "build:chrome": "webpack --env browser=chrome --env mode=production",
     "start:firefox": "webpack --env browser=firefox --watch",
-    "build:firefox": "export NODE_ENV=production && webpack --env browser=firefox --env mode=production",
-    "test": "jest --no-cache",
-    "version": "node sync-manifest.js && git add src/manifest-chrome.json && git add src/manifest-firefox.json"
+    "build:firefox": "webpack --env browser=firefox --env mode=production",
+    "test": "jest --no-cache"
   },
   "dependencies": {
     "@apollo/client": "^3.4.0-beta.4",
     "@fortawesome/fontawesome-svg-core": "^1.2.32",
     "@fortawesome/free-brands-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.1",
+    "@plussub/extension": "^5.2.5",
     "@plussub/srt-vtt-parser": "^1.0.6",
     "apollo": "^2.32.1",
     "ass-compiler": "^0.1.1",
@@ -37,23 +38,22 @@
     "@fortawesome/fontawesome-free": "^5.14.0",
     "@tailwindcss/forms": "^0.3.3",
     "@types/chart.js": "^2.9.29",
-    "@types/chrome": "^0.0.145",
+    "@types/chrome": "^0.0.159",
     "@types/firefox-webext-browser": "^82.0.0",
-    "@types/jest": "^26.0.16",
-    "@typescript-eslint/eslint-plugin": "^4.9.0",
-    "@typescript-eslint/parser": "^4.9.0",
+    "@types/jest": "^27.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
     "@vue/compiler-sfc": "^3.0.6",
     "autoprefixer": "^10.1.0",
     "copy-webpack-plugin": "^9.0.0",
-    "css-loader": "^5.0.1",
-    "eslint": "^7.14.0",
+    "css-loader": "^6.4.0",
+    "eslint": "^8.0.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-vue": "^7.2.0",
     "file-loader": "^6.2.0",
     "graphql-tag": "^2.11.0",
     "html-loader": "^2.1.2",
     "jest": "^27.0.4",
-    "jest-marbles": "^2.5.1",
     "postcss": "^8.2.0",
     "postcss-loader": "^6.1.0",
     "prettier": "^2.2.1",
@@ -63,6 +63,7 @@
     "ts-loader": "^9.2.3",
     "typescript": "^4.3.5",
     "url-loader": "^4.1.1",
+    "vue-eslint-parser": "^7.11.0",
     "vue-loader": "^16.5.0",
     "web-ext": "^6.1.0",
     "webpack": "^5.39.0",
