@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   key: 'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAr6+PlzeOK4z+p7QL/4NxAHoS6rhnNasLef6Dg6hp8E2/YQvKozVU4CFXnUR5sSmqJsVr8818+Q3vX4+zJbTa9S430HtudSqYGtpUyLzQTKmIGudXu/xNBS1cDRSEGqI+hKQBn3drNAd3zAnx8dHJBzUV5G6yJns2jYJ7GDFcDAK0N7C+F0WFPWCBhKAxH7/VRqzFiehP0Wr0NJN3hC2qX4KuJgiiAi0haeJ7ZmIWGCDgWM9zW6C2pPTdPbmVMQKvIKJzdKcj1CCVYrph0nqxNsDhhPxNd9HXshxquvskmQAdwcfW8kcWtU7/84Ib/+AvTLeVaXL3lmffZUnNOLfQKwIDAQAB'
 };
